Move XmlStyle.invoke function
