+++ viewdsl/src/main/java/splitties/viewdsl/core/styles/Styles.kt
@@ -15,11 +15,23 @@
  */
 package splitties.viewdsl.core.styles
 
+import android.content.Context
+import android.support.annotation.IdRes
+import android.support.annotation.StyleRes
 import android.view.View
+import splitties.viewdsl.core.NO_THEME
 import splitties.viewdsl.core.Style
+import splitties.viewdsl.core.styledView
+
+inline operator fun <reified V : View> XmlStyle<V>.invoke(
+        ctx: Context,
+        @IdRes id: Int = View.NO_ID,
+        @StyleRes theme: Int = NO_THEME,
+        initView: V.() -> Unit = {}
+): V = ctx.styledView(style = this, id = id, theme = theme, initView = initView)
 
 @Suppress("NOTHING_TO_INLINE")
-inline fun <V: View> V.apply(style: Style<V>): V {
+inline fun <V : View> V.apply(style: Style<V>): V {
     with(style) {
         applyStyle()
     }
